[T3CMS] Introduce project settings
To resolve some issues and enable customization, this change introduces
project settings and most importantly a way to enable / disable
the plugin functionality for a project.

The TYPO3CMSProjectComponent tries to detect a TYPO3 CMS project and
proposes to enable itself.

Followups should include enabling / disabling inspections and other
features based on this setting which should slightly improve performance
in projects that don't need the functionality.